+++ apps/react/demo/src/data/transformers.ts
@@ -1,9 +1,6 @@
-import type {
-  ExampleControlFieldAllTransformers,
-  ExampleSettingValue,
-} from "../ui/index.js";
+import type { ExampleSettingValue } from "../ui/index.js";
 
-const convertToPixels = (value: ExampleSettingValue) => `${value}px`;
+export const convertToPixels = (value: ExampleSettingValue) => `${value}px`;
 
 /**
  * Converts a value to the product of itself and the baseline height.
@@ -11,10 +8,10 @@ const convertToPixels = (value: ExampleSettingValue) => `${value}px`;
  * @param value The value to convert.
  * @returns CSS string representing the value multiplied by the baseline height.
  */
-const convertToBaselineMultiples = (value: ExampleSettingValue) =>
+export const convertToBaselineMultiples = (value: ExampleSettingValue) =>
   `calc(${value} * var(--baseline-height))`;
 
-const convertToRems = (value: ExampleSettingValue) => `${value}rem`;
+export const convertToRems = (value: ExampleSettingValue) => `${value}rem`;
 
 /*
   Because we are using the shadow DOM to isolate styles from the rest of the page,
@@ -23,22 +20,5 @@ const convertToRems = (value: ExampleSettingValue) => `${value}rem`;
   To get around this, the in-browser demo directly multiplies font-relative settings by the root font size.
   Then, when exporting, we can convert the values to `rem` units, to simplify the exported CSS.
  */
-const convertToFontSizeMultiples = (value: ExampleSettingValue) =>
+export const convertToFontSizeMultiples = (value: ExampleSettingValue) =>
   `calc(${value} * var(--font-size))`;
-
-/**
- * A pair of transformers for converting settings to functions of the font size in the demo,
- * and to `rem` units in the exported CSS.
- */
-const fontRelativeTransformers: ExampleControlFieldAllTransformers = {
-  demoTransformer: convertToFontSizeMultiples,
-  exportTransformer: convertToRems,
-};
-
-export default {
-  convertToPixels,
-  convertToRems,
-  convertToFontSizeMultiples,
-  convertToBaselineMultiples,
-  fontRelativeTransformers,
-};