+++ apps/react/demo/src/ui/Drawer/Drawer.stories.tsx
@@ -2,8 +2,9 @@
 
 import { Button } from "@canonical/react-ds-core";
 import type { Meta, StoryFn, StoryObj } from "@storybook/react";
-import { useState } from "react";
+import { useRef, useState } from "react";
 import Component from "./Drawer.js";
+import type { DrawerProps } from "./types.js";
 
 const meta = {
   title: "Drawer",
@@ -19,22 +20,67 @@ export default meta;
 */
 type Story = StoryObj<typeof meta>;
 
-export const Default: StoryFn<typeof Component> = () => {
-  const [isOpen, setIsOpen] = useState(false);
+const Example = ({
+  isOpenOverride = false,
+  onClose,
+  children = <span>Hello world!</span>,
+  ...props
+}: Partial<DrawerProps>) => {
+  const [isOpen, setIsOpen] = useState(isOpenOverride);
   return (
     <>
       <Button
         onClick={() => setIsOpen((curIsOpen) => !curIsOpen)}
         type="button"
-        label={"Toggle drawer"}
-      />
+      >
+        Toggle Drawer
+      </Button>
       <Component
         title="Drawer title"
         isOpenOverride={isOpen}
         onClose={() => setIsOpen(false)}
+        {...props}
       >
-        <span>Hello world!</span>
+        {children}
       </Component>
     </>
   );
 };
+
+export const Default: StoryFn<typeof Component> = () => {
+  return <Example />;
+};
+
+// The default example hides the drawer by default, so regression testing it is not helpful.
+Default.parameters = {
+  chromatic: {
+    disable: true,
+  },
+};
+
+export const Open: StoryFn<typeof Component> = () => (
+  <Example isOpenOverride={true} />
+);
+
+// Hide this story from the sidebar/storybook UI but keep it for visual regression testing
+Open.tags = ["!dev"];
+
+export const Overflow: StoryFn<typeof Component> = () => {
+  const dialogueRef = useRef<HTMLDialogElement>(null);
+
+  return (
+    <Example
+      style={{ overflow: "visible" }}
+      enableOverflow={true}
+      dialogueRef={dialogueRef}
+    >
+      <p style={{ border: "1px solid red", position: "relative", left: -200 }}>
+        I am overflowing the drawer. Clicking inside this element should not
+        cause the drawer to close, even if the click is outside of the contents
+        bounding box.
+      </p>
+    </Example>
+  );
+};
+
+Overflow.parameters = Default.parameters;