+++ apps/react/demo/src/ui/Drawer/useDrawer.ts
@@ -6,8 +6,11 @@ const useDrawer = ({
   onOpen,
   onClose,
   parent,
+  dialogueRef: dialogueRefProp,
 }: UseDrawerProps): UseDrawerResult => {
-  const dialogueRef = useRef<HTMLDialogElement>(null);
+  const dialogueRef = useRef<HTMLDialogElement | null>(
+    dialogueRefProp ? dialogueRefProp.current : null,
+  );
   const isSsr = useMemo(() => typeof window === "undefined", []);
   const parentElement = useMemo(
     () => (isSsr ? undefined : parent || document.body),
@@ -17,22 +20,36 @@ const useDrawer = ({
 
   const open = useCallback(
     (event?: Event) => {
-      if (!parentElement) return;
+      if (
+        isSsr ||
+        !parentElement ||
+        !dialogueRef.current ||
+        dialogueRef.current.open
+      )
+        return;
       parentElement.style.overflow = "hidden";
       setIsOpenInternal(true);
+      dialogueRef.current.showModal();
       onOpen?.(event);
     },
-    [parentElement, onOpen],
+    [isSsr, parentElement, onOpen],
   );
 
   const close = useCallback(
     (event?: Event) => {
-      if (!parentElement) return;
+      if (
+        isSsr ||
+        !parentElement ||
+        !dialogueRef.current ||
+        !dialogueRef.current.open
+      )
+        return;
       parentElement.style.overflow = "";
       setIsOpenInternal(false);
+      dialogueRef.current.close();
       onClose?.(event);
     },
-    [parentElement, onClose],
+    [isSsr, parentElement, onClose],
   );
 
   const toggle = useCallback(
@@ -72,21 +89,47 @@ const useDrawer = ({
 
   // Handle outside clicks
   useEffect(() => {
-    if (isSsr || !parentElement) return;
+    const dialogueNode = dialogueRef.current;
+    if (isSsr || !parentElement || !dialogueNode) return;
 
     const handleClickOutside = (event: MouseEvent) => {
+      /**
+       * The dialogue is opened with `showModal()`, which creates a full-screen backdrop.
+       * This causes *all* click events to be captured by the dialogue - so we can't just check if the event target is contained by the dialogue as that is always true.
+       * Instead, we can check if the click is inside the dialogue box itself.
+       */
+      const dialogueBbox = dialogueNode.getBoundingClientRect();
+      const clickIsInsideDialogueBox =
+        event.clientX >= dialogueBbox.left &&
+        event.clientX <= dialogueBbox.right &&
+        event.clientY >= dialogueBbox.top &&
+        event.clientY <= dialogueBbox.bottom;
+
+      /**
+       * The click target must be the dialogue node itself.
+       * This is true when the backdrop is clicked, or when some space in the dialogue that is not covered by
+       * some other element (like a button) is clicked.
+       */
+      const clickTargetIsDialogue = event.target === dialogueNode;
+
       if (
+        // Ignore clicks if the dialogue is closed
         isOpenInternal &&
-        !dialogueRef?.current?.contains(event.target as Node)
+        // Ignore clicks if the click is inside the dialogue box
+        !clickIsInsideDialogueBox &&
+        // Ignore clicks if the click target is some child of the dialogue.
+        // By combining the coordinates check and the target check, we can close if the outside space is clicked
+        // While allowing for cases where the clicked element bleeds outside the dialogue box (such as ctx menus or tooltips)
+        clickTargetIsDialogue
       ) {
         close(event);
       }
     };
 
-    document.addEventListener("click", handleClickOutside);
+    dialogueNode.addEventListener("click", handleClickOutside);
 
     return () => {
-      document.removeEventListener("click", handleClickOutside);
+      dialogueNode.removeEventListener("click", handleClickOutside);
     };
   }, [parentElement, isSsr, close, isOpenInternal]);