+++ apps/react/demo/src/ui/Showcase/common/Example/common/Controls/Controls.tsx
@@ -1,8 +1,8 @@
-import { useState } from "react";
+import { useCallback, useState } from "react";
 import type { ReactElement } from "react";
 import type { ControlsProps } from "./types.js";
 import "./styles.css";
-import { Button, TooltipArea } from "@canonical/react-ds-core";
+import { Button } from "@canonical/react-ds-core";
 import { Field } from "@canonical/react-ds-core-form";
 import { Drawer } from "ui/Drawer/index.js";
 import { useShowcaseContext } from "../../hooks/index.js";
@@ -17,88 +17,69 @@ const Controls = ({ id, className, style }: ControlsProps): ReactElement => {
     activateNextExample,
     copyOutput,
     resetActiveExample,
+    showBaselineGrid,
+    toggleShowBaselineGrid,
   } = useShowcaseContext();
 
   const [settingsOpen, setSettingsOpen] = useState(false);
+  const toggleSettingsOpen = useCallback(
+    () => setSettingsOpen((settingsOpen) => !settingsOpen),
+    [],
+  );
 
   return (
-    <>
-      <div
-        id={id}
-        className={[componentCssClassname, className].filter(Boolean).join(" ")}
-        style={style}
-      >
-        <div className="start">
-          {/*TODO use icon buttons when icon is implemented*/}
-          <Button
-            label="<"
-            type="button"
-            onClick={activatePrevExample}
-            aria-label="Previous example"
-          />
-          <Button
-            label=">"
-            type="button"
-            onClick={activateNextExample}
-            aria-label="Next example"
-          />
-          <span id="active-example-name" style={{ color: "white" }}>
-            {activeExample.name}&nbsp;
-            {/*Zindex is usd to draw the tooltip over the baseline overlay and the controls element. */}
-            <TooltipArea
-              Message={<span>{activeExample.description}</span>}
-              messageElementClassName="example-description"
-              maxWidth={"275px"}
-              zIndex={202}
-            >
-              {/*TODO replace this with an information icon when icon is implemented*/}
-              <i>(i)</i>
-            </TooltipArea>
-          </span>
-        </div>
-        <div className="end">
-          <Field
-            id="baseline-toggler"
-            inputType="checkbox"
-            name="showBaselineGrid"
-            label="Baseline grid"
-            isOptional={true}
-          />
-          <Button
-            label="Settings"
-            type="button"
-            onClick={() => setSettingsOpen(true)}
-          />
-          <Button
-            type="button"
-            label="Copy CSS"
-            disabled={!demoOutput?.css}
-            onClick={() => copyOutput("css")}
-          />
-        </div>
+    <div
+      id={id}
+      className={[componentCssClassname, className].filter(Boolean).join(" ")}
+      style={style}
+    >
+      <div>
+        {/*TODO use icon buttons when icon is implemented*/}
+        <Button
+          type="button"
+          onClick={activatePrevExample}
+          aria-label="Open previous example"
+        >
+          Prev
+        </Button>
+        <Button
+          type="button"
+          onClick={activateNextExample}
+          aria-label="Open next example"
+        >
+          Next
+        </Button>
       </div>
+
+      {/*
+        TODO we need a way to invert color themes once we implement theming
+      */}
+      <h4 id="active-example-name">{activeExample.name}</h4>
+
       <Drawer
+        id="showcase-settings-drawer"
         title={`${activeExample.name} settings`}
         isOpenOverride={settingsOpen}
         onClose={() => setSettingsOpen(false)}
-        contentsClassName="inputs-drawer-contents"
       >
         <Button
-          label={"Reset to defaults"}
           type={"button"}
           onClick={resetActiveExample}
-        />
+          aria-label="Reset this example's settings to their default values"
+        >
+          Reset to defaults
+        </Button>
 
         {activeExample.sections.map((section) => (
-          <div className="setting-category" key={section.title}>
+          <section className="setting-category" key={section.title}>
             <h4>{section.title}</h4>
-            <div className="inputs">
+            <div className="fields">
               {section.fields.map(
                 ({
                   name,
                   defaultValue,
                   demoTransformer,
-                  exportTransformer,
+                  transformer,
                   disabledOutputFormats,
                   ...fieldProps
                 }) => (
@@ -111,10 +92,35 @@ const Controls = ({ id, className, style }: ControlsProps): ReactElement => {
                 ),
               )}
             </div>
-          </div>
+          </section>
         ))}
       </Drawer>
-    </>
+      <div className="end">
+        <Button
+          type="button"
+          onClick={toggleShowBaselineGrid}
+          aria-label={`Set baseline visibility to ${!showBaselineGrid}`}
+        >
+          {showBaselineGrid ? "Hide" : "Show"} baseline
+        </Button>
+        <Button
+          type="button"
+          onClick={toggleSettingsOpen}
+          aria-label={`Set settings visibility to ${!settingsOpen}`}
+        >
+          {settingsOpen ? "Close" : "Open"} settings
+        </Button>
+        {/* TODO this should probably create a toast or notification that the CSS was copied successfully. */}
+        <Button
+          type="button"
+          disabled={!demoOutput?.css}
+          aria-label="Copy the current CSS variables to the clipboard"
+          onClick={() => copyOutput("css")}
+        >
+          Copy CSS
+        </Button>
+      </div>
+    </div>
   );
 };