+++ packages/react/ds-core-form/src/ui/Field/inputs/Text/types.ts
@@ -7,6 +7,12 @@ type NativeTextProps = React.InputHTMLAttributes<HTMLInputElement>;
 type AdditionalTextProps = {
   /* The type of input field Enum*/
   inputType?: TODONativeInputTypes;
+
+  /* The prefix element. Will be a child of `span.prefix` */
+  prefix?: React.ReactNode;
+
+  /* The suffix element. Will be a child of `span.suffix` */
+  suffix?: React.ReactNode;
 };
 
 export type TextProps = InputProps<NativeTextProps & AdditionalTextProps>;