+++ packages/react/ds-core-form/src/ui/constants.ts
@@ -1,0 +1,10 @@
+export enum states {
+  Danger = "danger",
+  Success = "success",
+  Idle = "idle",
+  Disabled = "disabled",
+}
+
+export enum mixins {
+  FormInput = "form-input",
+}